add product business rules

+++ backend/productservice/src/main/java/com/etiya/productservice/service/concretes/CampaignProductServiceImpl.java
@@ -1,10 +1,12 @@
 package com.etiya.productservice.service.concretes;
 
+import com.etiya.productservice.core.configuration.exceptions.type.BusinessException;
 import com.etiya.productservice.dto.campaignProduct.*;
 import com.etiya.productservice.entity.Attribute;
 import com.etiya.productservice.entity.CampaignProduct;
 import com.etiya.productservice.mapper.CampaignProductMapper;
 import com.etiya.productservice.repository.CampaignProductRepository;
+import com.etiya.productservice.rules.CampaignProductBusinessRules;
 import com.etiya.productservice.service.abstracts.CampaignProductService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -18,6 +20,8 @@ import java.util.UUID;
 public class CampaignProductServiceImpl implements CampaignProductService {
 
     private final CampaignProductRepository campaignProductRepository;
+
+    private final CampaignProductBusinessRules campaignProductBusinessRules;
 
     CampaignProductMapper campaignProductMapper = CampaignProductMapper.INSTANCE;
     @Override
@@ -33,21 +37,31 @@ public class CampaignProductServiceImpl implements CampaignProductService {
 
     @Override
     public CreateCampaignProductResponse create(CreateCampaignProductRequest request) {
+        campaignProductBusinessRules.checkIfProductExist(request.getProductId());
+        campaignProductBusinessRules.checkIfCampaignExist(request.getCampaignId());
         CampaignProduct campaignProduct =  campaignProductRepository.save(campaignProductMapper.campaignProductFromCreateRequest(request));
         return campaignProductMapper.campaignProductFromCreateResponse(campaignProduct);
     }
 
     @Override
     public UpdateCampaignProductResponse update(UpdateCampaignProductRequest request) {
+        campaignProductBusinessRules.checkIfCampaignProductExists(request.getId());
+        campaignProductBusinessRules.checkIfProductExist(request.getProductId());
+        campaignProductBusinessRules.checkIfCampaignExist(request.getCampaignId());
         CampaignProduct campaignProduct =  campaignProductRepository.save(campaignProductMapper.campaignProductFromUpdateRequest(request));
         return campaignProductMapper.campaignProductFromUpdateResponse(campaignProduct);
     }
 
     @Override
     public DeleteCampaignProductResponse delete(UUID id) {
-        campaignProductRepository.deleteById(id);
-        return campaignProductMapper.campaignProductFromDeleteResponse(campaignProductRepository.findById(id).orElseThrow());
-    }
+        Optional<CampaignProduct> campaignProduct = campaignProductRepository.findById(id);
 
+        if (campaignProduct.isPresent()) {
+            campaignProductRepository.deleteById(id);
+            return campaignProductMapper.campaignProductFromDeleteResponse(campaignProduct.get());
+        } else {
+            throw new BusinessException("No campaign product found to be deleted: " + id);
+        }
+    }
 
 }
